+++ packages/bpk-component-map/stories.js
@@ -20,7 +20,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { storiesOf } from '@storybook/react';
-import { action } from '@storybook/addon-actions';
+import { action } from 'bpk-storybook-utils';
 import BpkText from 'bpk-component-text';
 import { withRtlSupport } from 'bpk-component-icon';
 import LandmarkIconSm from 'bpk-component-icon/sm/landmark';