+++ packages/bpk-component-router-link/stories.js
@@ -18,7 +18,7 @@
 
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { action } from '@storybook/addon-actions';
+import { action } from 'bpk-storybook-utils';
 
 import BpkRouterLink from './index';